+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -23,7 +23,7 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.domain.Price;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.order.dto.request.OrderLineItemCreateRequest;
 import kitchenpos.order.dto.request.OrderCreateRequest;
 import kitchenpos.order.dto.request.OrderUpdateStatusRequest;
