+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -12,11 +12,12 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.dao.TableGroupRepository;
+import kitchenpos.tablegroup.repository.TableGroupRepository;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.request.TableGroupCreateRequest;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.dto.TableGroupCreateRequest;
+import kitchenpos.tablegroup.application.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
